+++ lib/elixir/src/elixir_dispatch.erl
@@ -381,7 +381,7 @@ in_erlang_functions() ->
     { byte_size, 1 },
     % { date, 0 },
     { exit, 1 },
-    { float, 1 },
+    % { float, 1 },
     { float_to_binary, 1 },
     { float_to_list, 1 },
     { hd, 1 },